+++ packages/react-widgets/src/models/utils.js
@@ -7,16 +7,30 @@ import {
 import { FullyQualifiedName } from './fqn';
 import { MAP_TYPES, API_VERSIONS } from '@carto/react-api';
 
+export { getSpatialFiltersResolution } from './spatialFiltersResolution';
+
 export function isRemoteCalculationSupported(props) {
   const { source } = props;
 
-  return (
-    source &&
-    source.type !== MAP_TYPES.TILESET &&
-    source.credentials.apiVersion !== API_VERSIONS.V2 &&
-    !(source.geoColumn && getSpatialIndexFromGeoColumn(source.geoColumn)) &&
-    source.provider !== 'databricks'
-  );
+  if (
+    !source ||
+    source.type === MAP_TYPES.TILESET ||
+    source.credentials.apiVersion === API_VERSIONS.V2 ||
+    source.provider === 'databricks'
+  ) {
+    return false;
+  }
+
+  const isDynamicSpatialIndex =
+    source.geoColumn && getSpatialIndexFromGeoColumn(source.geoColumn);
+  if (
+    isDynamicSpatialIndex &&
+    !source.dataResolution &&
+    !source.spatialFiltersResolution
+  ) {
+    return false;
+  }
+  return true;
 }
 
 export function wrapModelCall(props, fromLocal, fromRemote) {