+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- Remote calculation for dynamic spatial index sources [#908](https://github.com/CartoDB/carto-react/pull/908)
+
 ## 3.0.0
 
 ### 3.0.0-alpha.21 (2024-09-26)